[feature] #1179: Add revoke-permission-or-role instruction

[feature] #1179: Add revoke instruction.
